+++ projects/angular-sdk/src/components/cfar-offer-banner-large/cfar-offer-banner-large.component.ts
@@ -234,8 +234,8 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
           totalPrice: "71.96"
         },
         offerDescription: [
-          "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",
-          "Cancel and choose between a 80% refund of your flight base fare and taxes or 100% airline travel credit",
+          "Add the flexibility to cancel your flight for any reason up to 24 hours before departure",
+          "Cancel and get a refund of your flight base fare and taxes, excluding additional services (paid seats, paid bags...)",
           "Get instant resolution, no forms or claims required"
         ]
       },
@@ -309,8 +309,8 @@ export class CfarOfferBannerLargeComponent extends GlobalComponent implements On
           totalPrice: "71.96"
         },
         offerDescription: [
-          "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",
-          "Cancel and choose between a 100% refund of your flight base fare and taxes or 100% airline travel credit",
+          "Add the flexibility to cancel your flight for any reason up to 24 hours before departure",
+          "Cancel and get a refund of your flight base fare and taxes, excluding additional services (paid seats, paid bags...)",
           "Get instant resolution, no forms or claims required"
         ]
       }