+++ projects/angular-sdk/src/i18n/en.ts
@@ -254,7 +254,11 @@ export const locale: I18n = {
                 YE: "Yemen",
                 ZM: "Zambia",
                 ZW: "Zimbabwe"
-            }
+            },
+            BUTTON: {
+                CLOSE: 'Close'
+            },
+            TERMS_LINK_TEXT: "Terms & conditions apply"
         },
         CFAR_OFFER_BANNER: {
             TITLE: "Cancel your flight at any time, for any reason.",
@@ -275,7 +279,6 @@ export const locale: I18n = {
         },
         CFAR_OFFER_DIALOG: {
             TITLE: "Cancel For Any Reason",
-            POWERED_BY: "Powered by Hopper",
             VIEW_DETAILS: "View details",
             BUTTON: {
                 CANCEL: "Continue Without Coverage",
@@ -283,8 +286,7 @@ export const locale: I18n = {
             },
             NO_OFFERS: "No offer available",
             LOADING: "Loading offers in progress...",
-            TERMS_WARNING: "If you don't add this, your ticket is non-refundable.",
-            TERMS_LINK_TEXT: "Terms & Conditions"
+            TERMS_WARNING: "If you don't add this, your ticket is non-refundable."
         },
         CFAR_EXERCISE_DIALOG: {
             TITLE: "Cancel flight",
@@ -308,8 +310,7 @@ export const locale: I18n = {
                 CANCEL: "Cancel",
                 SUBMIT: "Yes, cancel flight",
                 BACK: "Back",
-                CONTINUE: "Continue",
-                CLOSE: "Close"
+                CONTINUE: "Continue"
             },
             NO_CONTRACT: "No contract available",
             LOADING: "Loading in progress...",