+++ projects/angular-sdk/src/i18n/i18n.interface.ts
@@ -252,7 +252,11 @@ export interface I18n {
                 YE: string;
                 ZM: string;
                 ZW: string;
-            }
+            },
+            BUTTON: {
+                CLOSE: string;
+            },
+            TERMS_LINK_TEXT: string;
         },
         CFAR_OFFER_BANNER: {
             TITLE: string;
@@ -273,7 +277,6 @@ export interface I18n {
         },
         CFAR_OFFER_DIALOG: {
             TITLE: string;
-            POWERED_BY: string;
             VIEW_DETAILS: string;
             BUTTON: {
                 CANCEL: string;
@@ -282,7 +285,6 @@ export interface I18n {
             NO_OFFERS: string;
             LOADING: string;
             TERMS_WARNING: string;
-            TERMS_LINK_TEXT: string;
         },
         CFAR_EXERCISE_DIALOG: {
             TITLE: string;
@@ -307,7 +309,6 @@ export interface I18n {
                 SUBMIT: string;
                 BACK: string;
                 CONTINUE: string;
-                CLOSE: string;
             },
             NO_CONTRACT: string;
             LOADING: string;