+++ src/app/pages/guides/cfar-offer-dialog/cfar-offer-dialog.component.ts
@@ -21,7 +21,7 @@ import { CfarContractCustomer } from "projects/angular-sdk/src/apis/hopper-cloud
 export class CfarOfferDialogPageComponent extends CommonGuidesComponent {
 
   public htmlCode: string = `
-    <button mat-flat-button color="primary" (click)="onOpenCfarOfferDialog()">
+    <button mat-flat-button color="primary" aria-label="Open CFAR offer dialog" (click)="onOpenCfarOfferDialog()">
       Open CFAR offer dialog
     </button>
   `;
@@ -120,7 +120,7 @@ export class CfarOfferDialogPageComponent extends CommonGuidesComponent {
         name: 'itineraries',
         description: `
           Itineraries are required in order to create new offers <br />
-          See <a target="_blank" href="https://airlines-api.staging.hopper.com/airline/v1.0/docs/index.html#operation/postCustomerCfar_offers">
+          See <a target="_blank" aria-describedby="api documentation" href="https://airlines-api.staging.hopper.com/airline/v1.0/docs/index.html#operation/postCustomerCfar_offers">
             API documentation
           </a> <br />
           You need to pass itineraries OR cfarOffers.
@@ -131,7 +131,7 @@ export class CfarOfferDialogPageComponent extends CommonGuidesComponent {
         name: 'cfarOffers',
         description: `
           CfarOffers are loaded directly into the component (without creating new offers) <br />
-          See <a target="_blank" href="https://airlines-api.staging.hopper.com/airline/v1.0/docs/index.html#operation/postCustomerCfar_offers">
+          See <a target="_blank" aria-describedby="api documentation" href="https://airlines-api.staging.hopper.com/airline/v1.0/docs/index.html#operation/postCustomerCfar_offers">
             API documentation
           </a> <br />
           You need to pass itineraries OR cfarOffers.